// apps/web/src/lib/api-client/models/safeUserProfileDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AppStart API
 * API REST do AppStart - Especificação OpenAPI 3.0
 * OpenAPI spec version: 1.0.0
 */
import type { SafeUserProfileDtoRole } from './safeUserProfileDtoRole';

export interface SafeUserProfileDto {
  id: string;
  name: string;
  email: string;
  role: SafeUserProfileDtoRole;
}